Anshul 21-06-2023

+++ main/java/com/mdms/loco/locouncleansed/model/LocoUncleansedDataAddNewLoco.java
@@ -12,7 +12,7 @@ import com.fasterxml.jackson.annotation.JsonFormat;
 public class LocoUncleansedDataAddNewLoco {
 	@Id
 	@Column(name = "loco_no")
-	private int loco_no;
+	private long loco_no;
 
 	@Column(name = "loco_permanent_domain")
 	private String loco_permanent_domain;
@@ -177,11 +177,11 @@ public class LocoUncleansedDataAddNewLoco {
 		this.loco_auxilary = loco_auxilary;
 	}
 
-	public int getLoco_no() {
+	public long getLoco_no() {
 		return loco_no;
 	}
 
-	public void setLoco_no(int loco_no) {
+	public void setLoco_no(long loco_no) {
 		this.loco_no = loco_no;
 	}
