+++ main/java/com/mdms/loco/locouncleansed/repository/LocoUncleansedDataRepository.java
@@ -50,10 +50,12 @@ public interface LocoUncleansedDataRepository extends CrudRepository<LocoUnclean
 			+ "loco_entry_date, record_status, status, user_id, txn_date,\r\n"
 			+ "	remarks, loco_flag, loco_receiving_date)\r\n" + 
 			"VALUES (?1, ?2, ?3, ?4, ?5, ?6, ?7, ?8, ?9, ?10, ?11, ?12, ?13, ?14,?15,?16,?17,?18,?19,?20,?21,?22,?23);", nativeQuery=true)
-	int saveDieselBoardZonalLocoData(int locoNo, String locoPermanentDomain, String locoType, String locoOwningZone,
+	int saveDieselBoardZonalLocoData(long locoNo, String locoPermanentDomain, String locoType, String locoOwningZone,
 			String locoOwningDivision, Date locoMfgDt, String locOwningShed, String locoManufacturer, String locoLeasetype,
 			long locoInitialCost, long locoPOHCost, String tractionCode, String gaugeType, Long locoHaulingPower,
 			String locoMfgCountry,Date locoEtryDt,String recordstatus,String status,String uid,Date txndate,String remarks,String locoflag, Date locoRecdDate);
+	
+	
 	
 
 	@Transactional
@@ -64,11 +66,11 @@ public interface LocoUncleansedDataRepository extends CrudRepository<LocoUnclean
 	int updateDieselBoardZonalRecord( String locoPermanentDomain, String locoType, String locoOwningZone,
 			String locoOwningDivision, Date locoMfgDt, String locOwningShed, Date recddt,  String locoLeasetype,
 			 long locoInitialCost, long locoPOHCost, String flagCode, String gaugeType, Long locoHaulingPower,
-			String locoMfgCountry,String recordstatus,String status,String uid, Date locotxndate,Date locoentrydate ,String remarks, int locoNo);
+			String locoMfgCountry,String recordstatus,String status,String uid, Date locotxndate,Date locoentrydate ,String remarks, long locoNo);
 	
 	
 	@Query(value="select loco_no from mdms_loco.loco_uncleansed_data where loco_no=?1", nativeQuery=true)
-	Integer checklocoNoExist(int locoNo);
+	int checklocoNoExist(long locoNo);
 	
 	
 	@Transactional
@@ -77,7 +79,7 @@ public interface LocoUncleansedDataRepository extends CrudRepository<LocoUnclean
 			" loco_no, loco_permanent_domain, loco_type,loco_owning_zone, loco_owning_division, loco_manufacturing_date,loco_owning_shed,"
 			+ "loco_lease_type, loco_initial_cost, loco_poh_cost,loco_receiving_date,loco_entry_date, record_status, status, user_id, txn_date,remarks, loco_flag)\r\n" + 
 			"VALUES (?1, ?2, ?3, ?4, ?5, ?6, ?7, ?8, ?9, ?10, ?11, ?12,?13,?14,?15,?16,?17,?18);", nativeQuery=true)
-	int saveElectricBoardZonalData( int locoNo, String locoPermanentDomain, String locoType,
+	int saveElectricBoardZonalData( long locoNo, String locoPermanentDomain, String locoType,
 			String locoOwningZone, String locoOwningDivision, Date locoMfgDt, String locOwningShed,
 			String locoLeasetype, long locoInitialCost, long locoPOHCost, Date locoRcvdDate,Date locoEtryDt,String recordstatus,String status,String uid,Date txndate,String remarks,String locoflag);
